Db encapsulation (#7)
Adopted the repository pattern and put all persistent items behind a set of protocols in the repository ns. Created a sql implementation that packages everything nicely into a record that extends the protocols. The nice thing about this is the db doesn't leak at all into the rest of the code. You just just use the repo api with the repo object.

Also modified the way bulk deletes happen. It is a bug if your pagination trigger elements are unstarred because if they get dropped you never see what's below that is starred. The best approach, esp. given the potentially destructive nature of this, is to just reset the visible tap stream. There are still some tbd design decisions on this for party mode as the reset event only goes to the owner of the deleted taps. Once any sort of sharing happens there will be consequences.