Adding channels to keg-party (#32)
This PR adds the idea of a channel to keg-party. A user (no matter many clients they have connected) is always in a single channel. I've considered other options, but they make no sense since most tapping happens in a client agnostic manner. This allows for partitioning of taps along conceptual lines (e.g. PRs/Issues) and allows for multiplayer/sharing in a channel as taps are currently public.

This PR fixes deletion of unstarred taps so that the deleted taps are those belonging to the current user in their current channel. Otherwise it was too destructive.

Note that there are still some small issues related to syncing of deletions and favorites now that we're in multiuser space, but this will be for the future to fix.

Fixes #16